Incremental linearization (#535)
* initial commit for incremental linearization


* fix build error


* fix build error


* merge two PLW into one


* add only a secant line by default for the first linearization


* devide SplitPoint into TangentPoint and SecantPoint


* clip use option


* support numWorkers


* add concat

* clang format

* new file uncommited

* clang format

* fix test

---------

+++ src/engine/IEngine.h
@@ -90,7 +90,7 @@ public:
     /*
       Solve the encoded query.
     */
-    virtual bool solve( unsigned timeoutInSeconds ) = 0;
+    virtual bool solve( double timeoutInSeconds ) = 0;
 
     /*
       Retrieve the exit code.
